feat: fase 1 - fix migrations and add sensor_readings table
- Fix decimal(5,2) for float columns in rooms, sensor_latest_data, sensor_logs
- Fix unsignedSmallInteger for hmis.port
- Fix unsignedInteger for sensors.modbus_address_*
- Add idx_status and idx_last_read_at indexes to sensor_latest_data
- Fix composite index idx_room_time(room_id, created_at) in sensor_logs
- Add sensor_readings migration with composite idx_sensor_time
- Add MigrationSchemaTest with 10 tests (54 assertions)
- Switch phpunit.xml to MySQL for accurate schema testing

+++ database/migrations/2026_03_03_000001_create_rooms_table.php
@@ -12,8 +12,8 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->string('location')->nullable();
-            $table->float('temp_max_limit')->default(25.0);
-            $table->float('hum_max_limit')->default(60.0);
+            $table->decimal('temp_max_limit', 5, 2)->default(25.00);
+            $table->decimal('hum_max_limit', 5, 2)->default(60.00);
             $table->timestamps();
         });
     }
